requirements.txt
asgiref==3.2.10
certifi==2020.12.5
chardet==4.0.0
cloudinary==1.24.0
dj-database-url==0.5.0
Django==3.1
django-chartjs==2.2.1
django-cloudinary-storage==0.3.0
django-crispy-forms==1.9.2
django-heroku==0.3.1
django-storages==1.11.1
django-utils-six==2.0
gunicorn==20.0.4
idna==2.10
Pillow==8.0.1
psycopg2==2.8.6
pytz==2020.4
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
